requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.3
Babel==2.3.4
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
boto==2.46.1
celery==4.0.2
certifi==2017.1.23
coreapi==2.3.0
coreschema==0.0.4
Django==1.9
django-celery==3.2.1
django-cors-headers==2.0.2
django-rest-swagger==2.1.2
django-storages==1.5.2
djangorestframework==3.6.2
flower==0.9.1
futures==3.0.5
itypes==1.1.0
Jinja2==2.9.5
json2html==1.1.1
kombu==4.0.2
Markdown==2.6.8
MarkupSafe==1.0
oauthlib==2.0.2
openapi-codec==1.3.1
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2016.10
PyYAML==3.12
requests==2.13.0
requests-oauthlib==0.8.0
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.2
uritemplate==3.0.0
vine==1.1.3
